trying things to connect to backend search endpoint
not going very well,
i am able to get a 200 response from the endpoint, but it seems to just send back the url. i don't know if this is a backend problem, like i think it should be returning a json like series of results but its just the url. and i cant even get that to display but i don't want to put too much time into it if its not even correct